Added pagination to build list. Added loading builds in pycli

+++ .internal/pycli/entry.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python3
 
+# cd .internal/pycli && nodemon --no-stdin --exec python3 entry.py
+
 import blessed
 import yaml
 import ruamel.yaml
@@ -13,6 +15,7 @@ hostUser = os.getenv('HOSTUSER')
 iotstackPwd = os.getenv('IOTSTACKPWD')
 hostAddress = os.getenv('HOSTSSH_ADDR')
 hostPort = os.getenv('HOSTSSH_PORT')
+doRemoteCheck = False # Don't commit if this is False
 
 print('blessed Version:', blessed.__version__)
 print('ruamel.yaml Version:', ruamel.yaml.__version__)
@@ -23,33 +26,46 @@ print('iotstackPwd: ', iotstackPwd)
 print('SSH hostAddress: ', hostAddress)
 print('SSH hostPort: ', hostPort)
 
-print('')
-try:
-  print('Checking connectivity to host...') 
-  touchRes = execSilent('touch ./.tmp/rtest.file')
-  touchRes = execSilent('echo "exec success" >> ./.tmp/rtest.file')
-  readRes = execSilent('cat ./.tmp/rtest.file')
-  rmRes = execSilent('rm ./.tmp/rtest.file')
+sshState = ''
 
-  if readRes == 'exec success':
-    print('Connection and remote command execution successful') 
-  else:
+for x in sys.argv[1:]:
+  if x == '--no-remote':
+    doRemoteCheck = False
+
+print('')
+if doRemoteCheck:
+  try:
+    print('Checking connectivity to host...') 
+    touchRes = execSilent('touch ./.tmp/rtest.file')
+    touchRes = execSilent('echo "exec success" >> ./.tmp/rtest.file')
+    readRes = execSilent('cat ./.tmp/rtest.file')
+    rmRes = execSilent('rm ./.tmp/rtest.file')
+    if readRes == 'exec success':
+      sshState = '--ssh'
+      print('Connection and remote command execution successful')
+    else:
+      sshState = '--no-ssh'
+      print('Error attempting to execute commands on the host. You may need to regenerate SSH keys by running:')
+      print('  ./menu.sh --run-env-setup')
+      print('')
+      print('Or configure SSH to use correct ports.')
+      input("Press Enter to continue to menu...")
+
+  except Exception:
     print('Error attempting to execute commands on the host. You may need to regenerate SSH keys by running:')
     print('  ./menu.sh --run-env-setup')
     print('')
     print('Or configure SSH to use correct ports.')
+    print('')
+    print('Error reported:')
+    print(sys.exc_info())
+    traceback.print_exc()
+    print('')
     input("Press Enter to continue to menu...")
+    print('')
+  else:
+    print("Skipping remote check")
 
-except Exception:
-  print('Error attempting to execute commands on the host. You may need to regenerate SSH keys by running:')
-  print('  ./menu.sh --run-env-setup')
-  print('')
-  print('Or configure SSH to use correct ports.')
-  print('')
-  print('Error reported:')
-  print(sys.exc_info())
-  traceback.print_exc()
-  print('')
-  input("Press Enter to continue to menu...")
+  print("Loading IOTstack PyCLI menu...")
 
-os.system('python menu_main.py')
+os.system('python3 menu_main.py {sshState}'.format(sshState=sshState))
